Add closeBeginTag(), isWriting(), write(), newbie() to HtmlVisitor and remove HtmlViewVisitor dependency of HtmlView.
